#1450: temperedlb: move files to reflect new name
